Add SetFontSize to modifier method

+++ backend/internal/dto/update_request.go
@@ -3,5 +3,6 @@ package dto
 // That's what I've got from the frontend
 type UpdateRequest struct {
 	LineSpacing *float64 `json:"lineSpacing"`
-	FontSize *int `json:"fontSize"`
+	// FontSize is float because user could need font size of 13.5 for example. So it's better to be float
+	FontSize *float64 `json:"fontSize"`
 }
